roki: use double instead of float for pokemon

+++ Roki.Core/Services/Database/Maps/Pokemon.cs
@@ -8,11 +8,11 @@ namespace Roki.Services.Database.Maps
         public int Num { get; set; }
         public string Species { get; set; }
         public List<string> Types { get; set; }
-        public Dictionary<string, float> GenderRatio { get; set; }
+        public Dictionary<string, double> GenderRatio { get; set; }
         public Dictionary<string, int> BaseStats { get; set; }
         public Dictionary<string, string> Abilities { get; set; }
-        public float Height { get; set; }
-        public float Weight { get; set; }
+        public double Height { get; set; }
+        public double Weight { get; set; }
         public string Color { get; set; }
         public string Prevo { get; set; }
         public List<string> Evos { get; set; }
@@ -34,7 +34,7 @@ namespace Roki.Services.Database.Maps
         public string Name { get; set; }
         public string Desc { get; set; }
         public string ShortDesc { get; set; }
-        public float Rating { get; set; }
+        public double Rating { get; set; }
     }
 
     public class Move
